Cleaning up graphbolt/src/cuda/common.h
